More tests for jabba
